+++ packages/soda-hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # soda-hooks
 
+## 6.12.1
+
+### Patch Changes
+
+- Updated dependencies
+    - deepsea-tools@5.39.0
+
 ## 6.12.0
 
 ### Minor Changes